% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abcam.neuronal.markers}
\alias{abcam.neuronal.markers}
\title{A dataframe containing the abcam neuronal marker genes  
abcam.neuronal.markers}
\format{A data frame.}
\source{
\url{http://www.abcam.com/neuroscience/neural-markers-guide}
}
\usage{
abcam.neuronal.markers
}
\description{
A dataframe containing the abcam neuronal marker genes  
abcam.neuronal.markers
}
\keyword{datasets}
