% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addEmbedding}
\alias{addEmbedding}
\title{addEmbedding}
\usage{
addEmbedding(object, df, reduction = "tsne")
}
\description{
Given a Seurat object and a data frame where the rows correspond to cells,
in the same order as in the Seurat object, add two columns giving coordinates
in a dimensionality reduced space.
}
\examples{
df <- data.frame(Cell = rownames(pbmc@meta.data),
                 Cluster = pbmc@meta.data$res.0.8)

addEmbedding(pbmc, df, reduction = "tsne")
}
\author{
Selin Jessa
}
