% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{addLabels}
\alias{addLabels}
\title{Add cluster labels to a tSNE ggplot2 plot}
\usage{
addLabels(centers, label_repel = FALSE, label_size = 4,
  label_short = FALSE)
}
\arguments{
\item{centers}{Data frame with at least three columns: "mean_tSNE_1", "mean_tSNE_2",
and "Cluster", as returned by \code{\link{clusterCenters}}}

\item{label_repel}{Logical, whether to plot cluster
labels repelled from the center, on a slightly transparent white background and
with an arrow pointing to the cluster center. If FALSE, simply plot the
cluster label at the cluster center. Default: TRUE.}

\item{label_size}{Numeric, controls the size of text labels. Default: 4.}

\item{label_short}{(Optional/Experimental!!) Logical, if TRUE, assumes clusters
(at seurat@ident) consist of a prefix and a suffix separated by a non-alpha
numeric character (\code{"[^[:alnum:]]+"}), and tries to separate these names
and only plot the prefix, for shorter labels and a cleaner plot. Default: FALSE.}
}
\description{
Add cluster labels to a tSNE ggplot2 plot
}
\author{
Selin Jessa
}
