% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{annotateMarkers}
\alias{annotateMarkers}
\title{Annotate cluster markers computed by Seurat}
\usage{
annotateMarkers(markers, species = "mm")
}
\arguments{
\item{markers}{A dataframe containing cluster markers as computed by
\link[Seurat]{FindAllMarkers}. Rownames should be gene symbols.}

\item{species}{String specifying the species, one of "mm" or "hg", used to determine the
annotation. Default: "mm".}
}
\description{
Annotate cluster markers computed by Seurat
}
\examples{
annotateMarkers(markers_pbmc, "hg")
}
\author{
Selin Jessa
}
