% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell.cycle.genes.whitfield.2002}
\alias{cell.cycle.genes.whitfield.2002}
\title{A dataframe containing the gene symbols for each phase of the cell cycle.
cell.cycle.genes.whitfield.2002}
\format{A data frame.}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/12058064}
}
\usage{
cell.cycle.genes.whitfield.2002
}
\description{
A dataframe containing the gene symbols for each phase of the cell cycle.
cell.cycle.genes.whitfield.2002
}
\keyword{datasets}
