% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCyclePlot.R
\name{cellCyclePlot}
\alias{cellCyclePlot}
\title{cellCyclePlot}
\usage{
cellCyclePlot(seurat, facets = TRUE, species = "m_musculus")
}
\arguments{
\item{seurat}{Seurat object.}

\item{facets}{Boolean to indicate whether a facet should be used for each cluster., or if all the clusters should be plotted together.}

\item{species}{m_musculus or h_sapiens.}
}
\value{
A ggplot2 object. A tSNE plot with the cell cycle plots
}
\description{
cellCyclePlot
}
\author{
Alexis Blanchet-Cohen
}
