% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{clusterCenters}
\alias{clusterCenters}
\title{clusterCenters}
\usage{
clusterCenters(seurat, reduction, dim1, dim2)
}
\arguments{
\item{seurat}{Seurat object, where dimensionality reduction has been applied,
i.e. (after applying Seurat::RunTSNE() to the object).}
}
\value{
Data frame with three columns: Cluster, mean_tSNE_1, and mean_tSNE_2
}
\description{
Get centers of clusters given a Seurat object, to use for labelling
in tSNE space. The cluster center is defined as the median X and Y coordinate
across cells in each cluster.
}
\examples{

clusterCenters(pbmc, reduction = "pca", dim1 = 1, dim2 = 3)
}
\author{
Selin Jessa
}
