% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_cycle.R
\name{computeCC}
\alias{computeCC}
\title{computeCC}
\usage{
computeCC(seurat, species = "mm")
}
\arguments{
\item{seurat}{Seurat object}

\item{species}{"mm" or "hg"}
}
\value{
A seurat object with scores in the \code{@meta.data} slot in two
new columns, "G1_S_score" and "G2_M_score"
}
\description{
Compute the cell cycle scores for a Seurat object, consisting of the mean expression,
in every cell, of a set of G1/S and G2/M phase markers.
}
\author{
Selin Jessa
}
