% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{constrainScale}
\alias{constrainScale}
\title{Constrain the scale of the plot to the dimensionality reduction limits}
\usage{
constrainScale(seurat, reduction = "tsne")
}
\arguments{
\item{seurat}{Seurat object for which a dimensionality reduction has been
computed (e.g. PCA or tSNE)}

\item{reduction}{String, corresponding to the dimensionality reduction to use.
Default: "tsne".}
}
\description{
Constrain the scale of the plot to the dimensionality reduction limits
}
\author{
Selin Jessa
}
