% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{convertGenesBetweenSpeciesHomologs}
\alias{convertGenesBetweenSpeciesHomologs}
\title{convertGenesBetweenSpeciesHomologs}
\usage{
convertGenesBetweenSpeciesHomologs(x, ...)
}
\arguments{
\item{x}{Character vector or matrix with mouse gene symbols to convert.
If a matrix, rownames should be gene symbols.}

\item{from}{One of "hg" or "mm"}

\item{to}{One of "hg" or "mm"}
}
\description{
convertGenesBetweenSpeciesHomologs
}
\examples{
convertGenesBetweenSpeciesHomologs(c("TUBB", "NES"), "hg", "mm")

mat1 <- matrix(seq(1, 15), nrow = 3)
rownames(mat1) <- c("Gfap", "Fabp7", "Foo")
convertGenesBetweenSpeciesHomologs(mat1, from = "mm")
}
\author{
adapted from Claudia Kleinman
}
