% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{convertGenesBetweenSymbolsAndID}
\alias{convertGenesBetweenSymbolsAndID}
\title{Convert between gene symbols and ENSEMBL gene IDs}
\usage{
convertGenesBetweenSymbolsAndID(x, sp = "hg", from = "gene_symbol")
}
\arguments{
\item{x}{Character vector of genes, or matrix (e.g. of counts)
where rownames are genes.}

\item{sp}{Character, one of "mm" or "hg", specifying the species for which
data is given in \code{mat}. Default: "hg".}

\item{from}{Character vector, one of "gene_symbol" or "ensembl_gene_id",
specifying in which format the genes in \code{x} are provided.}
}
\description{
Convert between gene symbols and ENSEMBL gene IDs
}
\author{
Selin Jessa
}
