% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{convertName}
\alias{convertName}
\title{convertName}
\usage{
convertName(symbol, from, to = NULL)
}
\arguments{
\item{symbol}{String or character vector of gene names}

\item{from}{One of "hg" or "mm"}

\item{to}{One of "hg" or "mm"}
}
\value{
Character vector of converted gene names
}
\description{
convertName
}
\examples{
convertName(c("TUBB", "NES"), "hg", "mm")
}
