% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{drLims}
\alias{drLims}
\title{Get the limits of a the first two dimensions in a dimensionality reduction}
\usage{
drLims(seurat, reduction = "tsne")
}
\arguments{
\item{seurat}{Seurat object for which a dimensionality reduction has been
computed (e.g. PCA or tSNE)}

\item{reduction}{String, corresponding to the dimensionality reduction to use.
Default: "tsne".}
}
\value{
A list with two elements: "xlim", which is a character vector of
the limits for the x-axis, and "ylim", correspondingly for the y-axis
}
\description{
When plotting an embedding, we may want to plot specific cells, but
constrain the scale to match plots of the whole dataset. Given a dim.
reduction, this function extracts the x and y limits to use for plotting.
}
\examples{
drLims(pbmc)
}
\author{
Selin Jessa
}
