% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{ensembl2symbols}
\alias{ensembl2symbols}
\title{Convert ENSEMBL gene IDs to gene symbols}
\usage{
ensembl2symbols(x, sp = "hg")
}
\arguments{
\item{x}{Character vector of genes, or matrix (e.g. of counts)
where rownames are genes.}

\item{sp}{Character, one of "mm" or "hg", specifying the species for which
data is given in \code{mat}. Default: "hg".}
}
\description{
Convert ENSEMBL gene IDs to gene symbols
}
\examples{
mat1 <- matrix(seq(1, 15), nrow = 3)
rownames(mat1) <- c("ENSMUSG00000020932", "ENSMUSG00000019874", "Foo")
ensembl2symbols(mat1)

ensembl2symbols(c("ENSMUSG00000020932", "ENSMUSG00000019874"), "mm")
}
\author{
Selin Jessa
}
