% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchData}
\alias{fetchData}
\title{fetchData}
\usage{
fetchData(seurat, genes, clusters = NULL, return_cell = FALSE,
  return_cluster = FALSE, scaled = FALSE)
}
\arguments{
\item{seurat}{Seurat object}

\item{genes}{Genes to filter}

\item{clusters}{(Optional) Vector, include only cells with these identities
(e.g. cluster assignments). Searches in seurat@ident.}

\item{return_cell}{Logical, whether or not to include a column with the cell ID.
Default: FALSE}

\item{return_cluster}{Logical, whether or not to include a column with the cluster.
Default: FALSE}

\item{scaled}{Logical, whether to fetch scaled data from seurat@scale.data.
Default: FALSE.}
}
\value{
Expression matrix for genes specified
}
\description{
Subset the seurat@data matrix by gene and cluster.
Similar to Seurat::FetchData except it doesn't thrown an error if a gene
is not found in the data, and is more limited.
}
\examples{
fetchData(pbmc, c("IL32", "MS4A1"))
fetchData(pbmc, c("IL32"), c(1, 2))
fetchData(pbmc, c("IL32", "MS4A1"), c(1, 2), return_cluster = TRUE, return_cell = TRUE)
}
\author{
Selin Jessa
}
