% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findGenes}
\alias{findGenes}
\title{findGenes}
\usage{
findGenes(seurat, genes)
}
\arguments{
\item{seurat}{Seurat object}

\item{genes}{Character vector of genes}
}
\value{
A named list with two elements: "detected" and "undetected"
each storing character vectors with the genes in each category
}
\description{
Given a set of genes, find the ones which are detected in the sample,
and which are not.
}
\examples{
find_out <- findGenes(pbmc, c("IL32", "CD79B", "foo"))
find_out$detected
find_out$undetected
}
\author{
Selin Jessa
}
