% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{geneName}
\alias{geneName}
\title{geneName}
\usage{
geneName(symbols, species = c("hg", "mm"))
}
\arguments{
\item{symbols}{String, or character vector of gene names}

\item{species}{Species for which \code{symbols} are given}
}
\value{
Rows of annotation corresponding to the symbols
}
\description{
Find the proper capitalization (which is species dependent) for genes
TODO: return object does not preserve order of original list
}
\examples{
geneName("TUBB", species = "hg")
}
