% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerSelection.R
\name{get.AllMarkers}
\alias{get.AllMarkers}
\title{Title}
\usage{
get.AllMarkers(seurat, res, output.graphs = FALSE,
  output.table = FALSE, n_genes = 2, topn_markers = 10,
  graph.name = "Feature plot for cluster ",
  table.name = "Table of computed confidence values for resolution: ",
  n.trees = 0, dims.use = 1:20)
}
\arguments{
\item{seurat}{The Seurat Object}

\item{res}{The resolution of the clustering}

\item{output.graphs}{Indicates whether or not you wish to output the feature plots of the markers found to a pdf file}

\item{output.table}{Indicates whether or not you wish to output the table of markers to a pdf file}

\item{n_genes}{The number of genes used in each decision tree}

\item{topn_markers}{The number of markers you wish to return for each cluster}

\item{graph.name}{The name of the feature plot if one wishes to output them to a pdf file}

\item{table.name}{The name of the list of markers if one wishes to output it into a pdf file}

\item{n.trees}{The number of trees one wishes to construct for each cluster, if zero, then the value will be determined automatically as the number of most variable genes indicated in the seurat object}

\item{dims.use}{A vector that represents the PCs used in the clustering}
}
\value{
A dataframe that contains the list of markers and their associated attributes
}
\description{
Finds the markers for all of the clusters for a given resolution of the k-nn clustering
}
\examples{
marker_list<-get.AllMarkers(pbmc, res = 1,n.trees=50)
### If you wish to use your own clustering solution that has been previously prepaired see get.AllMarkers2
}
