% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerSelection.R
\name{get.AllMarkers2}
\alias{get.AllMarkers2}
\title{Title}
\usage{
get.AllMarkers2(seurat, labels, output.graphs = FALSE,
  output.table = FALSE, n_genes = 2, topn_markers = 10,
  graph.name = "Feature plot for cluster ",
  table.name = "Table of computed confidence values for resolution: ",
  n.trees = 0)
}
\arguments{
\item{seurat}{The Seurat Object}

\item{labels}{A dataframe that contains labels for the clustering solution. The dataframe should only have 1 column}

\item{output.graphs}{Indicates whether or not you wish to output the feature plots of the markers found to a pdf file}

\item{output.table}{Indicates whether or not you wish to output the table of markers to a pdf file}

\item{n_genes}{The number of genes used in each decision tree}

\item{topn_markers}{An integer indicating how many markers you wish to return for each cluster}

\item{graph.name}{The name of the feature plot if one wishes to output them to a pdf file}

\item{table.name}{The name of the list of markers if one wishes to output it into a pdf file}

\item{n.trees}{The number of trees one wishes to construct for each cluster, if zero, then the value will be determined automatically as the number of most variable genes indicated in the seurat object}
}
\value{
A dataframe that contains the list of markers and their associated attributes
}
\description{
Title
}
\examples{
marker_list<-get.AllMarkers2(pbmc, labels = data.frame(as.numeric(pbmc@meta.data$res.1)), n.trees = 50)

}
