% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{getClusterColours}
\alias{getClusterColours}
\title{Get a vector of cluster colours, optionally highlighting select clusters}
\usage{
getClusterColours(seurat, clusters = NULL, original_colours = NULL,
  default_colour = "gray80")
}
\arguments{
\item{seurat}{Seurat object}

\item{clusters}{Vector of one or more clusters to highlight, matching the levels at
\code{levels(seurat@ident)}. If "none", returns \code{default_colour}
for every clusters. Default: all clusters, obtained from \code{levels(seurat@ident)}.}

\item{original_colours}{(Optional) Vector of colours to use. Either one colour
per cluster, in the order of \code{levels(seurat@ident)}, or one colour per
cluster passed to \code{clusters}, in the other they were provided.
Default: default ggplot2 colours used by Seurat.}

\item{default_colour}{Colour to use for non-highlighted clusters
Default: gray80 (light grey).}
}
\value{
Named character vector
}
\description{
Given a seurat object, get a named character vector of cluster colours, where
names are cluster names (coresponding to \code{levels(seurat@ident)}), and
values are hex codes of the colours, either the default colours from Seurat,
or colours you specify. This is trivial to make this yourself,
so this is used as a utility function for retaining the colours of
clusters you want to highlight, and setting the colours of all other clusters
to grey, or another default, non-intrusive colour.
}
\examples{

# Trivial: get named character vector with default colours
getClusterColours(pbmc)

# Highlight clusters 2 and 3
getClusterColours(pbmc, clusters = c(2, 3))

# Highlight clusters 2 and 3, set all other cluster colours to white
getClusterColours(pbmc, clusters = c(2, 3), default_colour = "white")
}
\author{
Selin Jessa
}
