% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getVarianceExplained}
\alias{getVarianceExplained}
\title{getVarianceExplained}
\usage{
getVarianceExplained(seurat, n = 10)
}
\arguments{
\item{seurat}{Seurat object}

\item{n}{Numeric, number of PCs for which variance should be reported.
Default: 10}
}
\value{
List with two vectors, "percent.var.explained" and "cumulative.var.explained",
reported for the first \code{n} PCs
}
\description{
Compute variance explained by PCA, given a Seurat object for which the PCA
dim. reduction has been calculated
}
\examples{
getVarianceExplained(pbmc, n = 5)

}
\author{
Adapted from Yang Yang
}
