% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{ggColours}
\alias{ggColours}
\alias{ggColors}
\title{Get default ggplot2/Seurat colours}
\usage{
ggColours(n)
}
\arguments{
\item{n}{Number of colours to return}
}
\value{
Named character vector, where names are the names of clusters, from
0 to n-1, and values are the hex codes for the colours.
}
\description{
Get evenly spaced colours from around the colour wheel, which are the default
colours assigned to clusters by Seurat. The output of this function can be
passed to the \code{scale_colour_manual()} and \code{scale_fill_manual()} functions
from ggplot2, as the \code{values} argument. (\code{\link{ggColors}} points
to this function.)
}
\examples{

n_clust <- 5
ggColours(n_clust)

}
\references{
https://stackoverflow.com/a/8197703
}
