% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparisons.R
\name{heatmapPercentMarkerOverlap}
\alias{heatmapPercentMarkerOverlap}
\title{heatmapPercentMarkerOverlap}
\usage{
heatmapPercentMarkerOverlap(markers1, markers2, s1_name = NULL,
  s2_name = NULL, marker_col = "gene")
}
\arguments{
\item{markers1}{Data frame of markers for one dataset as returned by Seurat::FindAllMarkers()}

\item{markers2}{Data frame of markers for second dataset as returned by Seurat::FindAllMarkers()}

\item{s1_name}{(Optional) string giving the name or ID of sample 1 (used for axis labels)}

\item{s2_name}{(Optional) string giving the name or ID of sample 2 (used for axis labels)}

\item{marker_col}{String specifying the column in the markers data frames which
specifies the cluster. By default, Seurat calls this "gene" (the default here); in the pipeline,
it may be called "external_gene_name".}
}
\description{
Generate a heatmap where rows correspond to clusters from one dataset and
columns correspond to the clusters from the second dataset. The value at [i, j]
is the minimum of the percent overlap between markers for Cluster i in dataset 1
and Cluster j in dataset 2.
}
\examples{
# Here, we compute the heatmap for one dataset (pbmc) with itself
heatmapPercentMarkerOverlap(markers_pbmc, markers_pbmc)
}
\author{
Selin Jessa
}
