% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{hg2mm}
\alias{hg2mm}
\title{hg2mm}
\usage{
hg2mm(x, ...)
}
\arguments{
\item{x}{Character vector or matrix with human gene symbols to convert.
If character, provide either individual strings or one character vector.
If a matrix, rownames should be gene symbols.}
}
\description{
Convert genes from human to mouse, wrapper around \code{\link{convertGenesBetweenSpeciesHomologs}}.
For vectors, the output is \strong{not order preserving}.
For matrices, order is preserved, but \strong{genes where the homolog is not
found are silently dropped}.
}
\examples{
hg2mm("TUBB", "NES")
hg2mm(c("TUBB", "NES"))
}
\author{
Selin Jessa, adapted from Claudia Kleinman
}
