% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{highlight}
\alias{highlight}
\title{Highlight select clusters on a tSNE plot}
\usage{
highlight(seurat, clusters, original_colours = NULL,
  default_colour = "gray80", ...)
}
\arguments{
\item{seurat}{Seurat object, where Seurat::RunTSNE() has been applied}

\item{clusters}{Vector of one or more clusters to highlight, matching the levels at
\code{levels(seurat@ident)}. If "none", all clusters are coloured by \code{default_colour}.}

\item{original_colours}{(Optional) Vector of colours to use. Either one colour
per cluster, in the order of \code{levels(seurat@ident)}, or one colour per
cluster passed to \code{clusters}, in the other they were provided.
Default: default ggplot2 colours used by Seurat.}

\item{default_colour}{Colour to use for non-highlighted clusters. Default: gray80
(light grey).}

\item{...}{(Optional) Other arguments passed to \code{\link{tsne}}.}
}
\value{
A ggplot2 object
}
\description{
Plot the tSNE embedding for the dataset, with cells from select clusters coloured
by either their original colours or provided colours, and cells from all
other clusters in another (non-intrusive) colour.
}
\examples{
# Highlight cluster 3 on the tSNE plot
highlight(pbmc, 3)

# Pass additional arguments to cytokit::tsne
highlight(pbmc, c(2, 3), label = FALSE, title = "Test highlight")

# Change default colour
highlight(pbmc, c(2, 3), default_colour = "lightblue")

# Specify the colours to highlight the clusters with
highlight(pbmc, c(2, 3), c("red", "blue"))
}
\author{
Selin Jessa
}
