% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_correlation.R
\name{meanClusterExpression}
\alias{meanClusterExpression}
\title{meanClusterExpression}
\usage{
meanClusterExpression(seurat, genes = NULL)
}
\arguments{
\item{seurat}{Seurat object}

\item{genes}{(Optional) Character vector specifying genes for which means
should be calculated. If any provided genes are not detected, they are
silently dropped. Default: all detected genes, i.e. \code{rownames(seurat@data)}.}
}
\value{
A gene x cluster matrix, where each column contains the mean
expression of each gene over all the cells in that cluster
}
\description{
This function computes the mean expression of each gene in each cluster
}
\examples{
meanClusterExpression(pbmc, c("IL32", "CD2"))
}
\author{
Selin Jessa
}
