% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{meanGeneExpression}
\alias{meanGeneExpression}
\title{meanGeneExpression}
\usage{
meanGeneExpression(seurat, genes)
}
\arguments{
\item{seurat}{Seurat object}

\item{genes}{String or character vector specifying gene(s) to use}
}
\value{
Data frame with two columns: "Cell" which specifies the cell ID
obtained from \code{colnames(seurat@data)}, and "Mean_marker_expression"
which is the expression value for the marker gene, or mean if multiple
genes were provided.
}
\description{
Calculate, for each cell, the mean expression of the given set of marker genes.
}
\examples{
meanGeneExpression(pbmc, genes = c("IL32", "MS4A1"))
}
\author{
Selin Jessa
}
