% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparisons.R
\name{meanMarkerExprByCluster}
\alias{meanMarkerExprByCluster}
\title{meanMarkerExprByCluster}
\usage{
meanMarkerExprByCluster(seurat, markers, marker_col = "gene")
}
\value{
Data frame where the first two columns are "Cell" and "Cluster",
and all remaining columns give the mean expression of markers identified
for each cluster
}
\description{
Compute the mean expression of markers for each cluster in one dataset,
in each cluster of another dataset. Used for \code{\link{markerViolinPlot}}
}
\author{
Selin Jessa
}
