% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons_with_markers.R
\name{meanMarkerExpressionPerCell}
\alias{meanMarkerExpressionPerCell}
\title{meanMarkerExpression}
\usage{
meanMarkerExpressionPerCell(seurat, markers, marker_col = "gene")
}
\arguments{
\item{seurat}{Seurat object, whose expression values will be used}

\item{markers}{Markers data frame, for the same or another Seurat object. The
expression of markers in each cluster in this data frame will be calculated
for each cluster in \code{seurat}.}
}
\value{
Data frame where the first two columns are "Cell" and "Cluster" (from
\code{seurat}),
and all remaining columns give the mean expression of markers identified
for each cluster in \code{markers}
}
\description{
This function computes each cell's mean expression of each cluster's
markers. Used for \code{\link{markerViolinPlot}}.
}
\examples{

library(dplyr)

# Using the same sample's marker:
meanMarkerExpressionPerCell(pbmc, markers_pbmc, "gene")

# Change the name of the clusters in the markers df, as if it were
# from a different sample where the clusters are A, B, C, D:
markers2 <- mutate(markers_pbmc, cluster = recode(
    cluster, `0` = "A", `1` = "B", `2` = "C", `3` = "D"))

meanMarkerExpressionPerCell(pbmc, markers2, "gene")
}
\author{
Selin Jessa
}
