% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{noLegend}
\alias{noLegend}
\title{Remove the legend in a ggplot}
\usage{
noLegend()
}
\value{
A theme element to hide legend
}
\description{
Remove the legend in a ggplot
}
\examples{
# gg <- mpg \%>\%
# filter(class \%in\% c("compact", "suv")) \%>\%
# ggplot(aes(x = displ, y = hwy)) +
# geom_point(aes(colour = factor(year))) +
# facet_wrap(~ class, ncol = 2)

# gg
# gg + noLegend()
}
\author{
Selin Jessa
}
