% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons_with_markers.R
\name{pairwiseVln}
\alias{pairwiseVln}
\alias{markerViolinPlot}
\title{pairwiseVln}
\usage{
pairwiseVln(seurat1, markers, seurat2,
  sample_names = c(seurat1@project.name, seurat2@project.name),
  marker_col = "gene")
}
\arguments{
\item{seurat1}{Seurat object for which to plot expression}

\item{markers}{Data frame as returned by Seurat::FindAllMarkers()}

\item{seurat2}{Seurat object from which cluster markers (\code{markers}) were defined}

\item{sample_names}{(Optional) Character vector giving the names or ID of
the two samples, used for axis labels, etc.
Default: c(seurat1@project.name, seurat2@project.name)}

\item{marker_col}{String specifying the column in the \code{markers} data frames which
specifies the cluster. By default, Seurat calls this "gene" (the default here);
in objects produced by the lab's pipeline, it may be called "external_gene_name".}
}
\value{
A ggplot object
}
\description{
Make a grid of violin plots with the mean expression of each cluster, for markers
dataset 1 is the sample whose expression values will be plot, and dataset 2
is the sample whose cluster markers are used.
}
\examples{
pairwiseVln(pbmc, markers_pbmc, pbmc)
}
\author{
adapted from Alexis Blanchet-Cohen
}
