% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{percentilesMarkerExpression}
\alias{percentilesMarkerExpression}
\title{perecentilesMarkerExpression}
\usage{
percentilesMarkerExpression(object, markers)
}
\arguments{
\item{object}{Seurat object}

\item{markers}{String or character vector specifying gene(s) to use}
}
\value{
Data frame with three columns: "Cell" which specifies the cell ID
obtained from \code{colnames(object@data)}, "Percentile" giving the
percentile expression for each cell, and
"Gradient_group" which gives the ordering that should be used to assign
colours by gradient.
}
\description{
perecentilesMarkerExpression
}
\author{
adapted from code from Alexis Blanchet-Cohen
}
