% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparisons.R
\name{percentMarkerOverlap}
\alias{percentMarkerOverlap}
\title{percentMarkerOverlap}
\usage{
percentMarkerOverlap(markers1, markers2, marker_col = "gene")
}
\arguments{
\item{markers1}{Data frame of markers for one dataset as returned by Seurat::FindAllMarkers()}

\item{markers2}{Data frame of markers for second dataset as returned by Seurat::FindAllMarkers()}

\item{marker_col}{String specifying the column in the markers data frames which
specifies the cluster. By default, Seurat calls this "gene" (the default here); in the pipeline,
it may be called "external_gene_name".}
}
\value{
Data frame with three columns: "s1_cluster" giving the cluster for
dataset 1, "s2_cluster" giving the cluster for dataset 2, and "marker_overlap"
which gives the computed min percent overlap of markers between the two clusters
}
\description{
percentMarkerOverlap
}
\examples{
percentMarkerOverlap(markers_pbmc, markers_pbmc)
}
\author{
Selin Jessa
}
