% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerSelection.R
\name{print_table}
\alias{print_table}
\title{Title}
\usage{
print_table(file.name, sga_hits)
}
\arguments{
\item{file.name}{The file name of the output file}

\item{sga_hits}{The data Frame to be printed}
}
\value{
No return value
}
\description{
Outputs a dataframe into a pdf file. Useful for very long tables that spans multiple pages
}
