% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{rotateX}
\alias{rotateX}
\title{Rotate the x axis labels in a ggplot}
\usage{
rotateX(angle = 90)
}
\arguments{
\item{angle}{Integer, value in degrees to rotate labels. Default: 90.}
}
\value{
A theme element to rotate labels
}
\description{
Rotate the x axis labels in a ggplot
}
\examples{
# gg <- mpg \%>\%
# filter(class \%in\% c("compact", "suv")) \%>\%
# ggplot(aes(x = displ, y = hwy)) +
# geom_point(aes(colour = factor(year))) +
# facet_wrap(~ class, ncol = 2)

# gg
# gg + rotateX()
}
\author{
Selin Jessa
}
