% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerSelection.R
\name{select_best_markers3}
\alias{select_best_markers3}
\title{Title}
\usage{
select_best_markers3(df, cluster_index, cluster_i, n.gene = 2,
  n.trees = 0)
}
\arguments{
\item{df}{A dataframe represents the single cell data matrix, with the genes as the features and cells as the rows. NAs are not allowed}

\item{cluster_index}{A one column dataframe that contains the cluster labels assigned to each cell.}

\item{cluster_i}{The cluster which we wish to extract the markers for}

\item{n.gene}{The number of genes used in each decision tree}

\item{n.trees}{The number of trees one wishes to construct for each cluster, if zero, then the value will be determined automatically as the number of most variable genes indicated in the seurat object}
}
\value{
The dataframe with the list of markers
}
\description{
This function finds the gene markers for a given cluster.
}
\examples{
### Suppose you want to extract the markers for cluster 0
markers<-select_best_markers3(t(pbmc@scale.data), cluster_index = data.frame(as.numeric(pbmc@meta.data$res.1)), cluster_i = 0)
### Usually this function is only used if one only wishes to extract markers for a single cluster and don't want to compute the rest to save computation time
### For getting markers for all the clusters, see get.AllMarkers and get.AllMarkers2
}
