% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkerSelection.R
\name{selectMarkersRF}
\alias{selectMarkersRF}
\title{Title}
\usage{
selectMarkersRF(df, cluster_index, cluster_i, n_genes = 2,
  n_trees = NCOL(df) * 2, reduced_form = TRUE)
}
\arguments{
\item{df}{A dataframe represents the single cell data matrix, with the genes as the features and cells as the rows. NAs are not allowed}

\item{cluster_index}{A one column dataframe that contains the cluster labels assigned to each cell.}

\item{cluster_i}{The cluster which we wish to extract the markers for}

\item{n_genes}{The number of genes used in each decision tree}

\item{n_trees}{The number of trees one wishes to construct for each cluster, default is the n_genes times the number of features in the data matrix}

\item{reduced_form}{Indicates whether or not the output should be in its reduced form. The full form contains the recall, percision and 1/error rate at each iteration.
Whereas the reduced form only contains the final recall, percision, and 1/error rate}
}
\value{
The dataframe with the list of markers
}
\description{
This function finds the gene markers for a given cluster.
}
\examples{
### Suppose you want to extract the markers for cluster 0
df <- t(pbmc@scale.data)
labels <- data.frame(as.numeric(pbmc@meta.data$res.1))
markers<-selectMarkersRF(df, cluster_index = labels, cluster_i = 0)
}
