% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R
\name{seurat_to_monocle}
\alias{seurat_to_monocle}
\title{Converts a Seurat object to a Monocle object.}
\usage{
seurat_to_monocle(seurat, selected_clusters)
}
\arguments{
\item{seurat}{Seurat object to convert to a Monocle (CellDataSet) object.}

\item{selected_clusters}{Selected clusters in Seurat object. Full names of clusters must be given.}
}
\value{
A CellDataSet object
}
\description{
Full name of selected clusters must be given.
In your script, you must load the monocle library, before calling this function.
}
\author{
Alexis Blanchet-Cohen
}
