% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{symbols2ensembl}
\alias{symbols2ensembl}
\title{Convert gene symbols to ENSEMBL gene IDs}
\usage{
symbols2ensembl(x, sp = "hg")
}
\arguments{
\item{x}{Character vector of genes, or matrix (e.g. of counts)
where rownames are genes.}

\item{sp}{Character, one of "mm" or "hg", specifying the species for which
data is given in \code{mat}. Default: "hg".}
}
\description{
Convert gene symbols to ENSEMBL gene IDs
}
\examples{
mat1 <- matrix(seq(1, 15), nrow = 3)
rownames(mat1) <- c("Gfap", "Fabp7", "Foo")
symbols2ensembl(mat1)

symbols2ensembl(c("Gfap", "Fabp7"), "mm")
}
\author{
Selin Jessa
}
