% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_min}
\alias{theme_min}
\title{theme_min}
\usage{
theme_min(base_size = 11, base_family = "")
}
\description{
A clean theme for ggplot2
}
\references{
https://github.com/sjessa/ggmin
}
\author{
Selin Jessa
}
