% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tsne}
\alias{tsne}
\title{tsne}
\usage{
tsne(seurat, label = TRUE, point_size = 0.6, alpha = 0.8,
  legend = FALSE, label_repel = TRUE)
}
\arguments{
\item{seurat}{Seurat object, where Seurat::RunTSNE() has been applied}

\item{label}{Logical, whether to plot cluster labels. Default: TRUE}

\item{point_size}{Numeric, size of points in scatter plot. Default: 0.6}

\item{alpha}{Numeric, fixed alpha value for points: Default: 0.8}

\item{legend}{Logical, whether to plot legend.}

\item{label_repel}{Logical, if \code{label} is TRUE, whether to plot cluster
labels repelled from the center, on a slightly transparent white background and
with an arrow pointing to the cluster center. If FALSE, simply plot the
cluster label at the cluster center. Default: TRUE.}
}
\value{
A ggplot2 object
}
\description{
Plot a coloured, labelled tSNE plot for a datasets, akin to Seurat::TSNEPlot()
}
\examples{
tsne(pbmc)
}
\author{
Selin Jessa
}
