% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle.R
\name{tsne_plot_clusters_highlighted}
\alias{tsne_plot_clusters_highlighted}
\title{Draws a tSNE plot highlighting the selected clusters.}
\usage{
tsne_plot_clusters_highlighted(seurat, selected_clusters, legend = TRUE)
}
\arguments{
\item{seurat}{Seurat object to convert to a Monocle (CellDataSet) object.}

\item{selected_clusters}{Selected clusters in Seurat object. Full names of clusters must be given.}

\item{legend.}{Boolean. Display legend or not.}
}
\value{
A ggplot2 object.
}
\description{
Draws a tSNE plot highlighting the selected clusters.
}
\author{
Alexis Blanchet-Cohen
}
