% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsneByPercentileMarkerExpressionAlexis.R
\name{tsneByPercentileMarkerExpressionAlexis}
\alias{tsneByPercentileMarkerExpressionAlexis}
\title{cellCyclePlot}
\usage{
tsneByPercentileMarkerExpressionAlexis(seurat, genes,
  legend = "percentiles")
}
\arguments{
\item{seurat}{Seurat object.}

\item{genes}{Genes to plot.}

\item{legend}{"percentiles", "expression" or "none"}
}
\value{
A ggplot2 object. A tSNE plot with the cell cycle plots
}
\description{
cellCyclePlot
}
\author{
Alexis Blanchet-Cohen
}
