% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vlnGrid}
\alias{vlnGrid}
\title{vlnGrid}
\usage{
vlnGrid(seurat, genes)
}
\arguments{
\item{seurat}{Seurat object}

\item{genes}{Genes to plot violins for}
}
\value{
A ggplot2 object
}
\description{
A grid of small violin plots, one plot per gene per cluster, similar to
Figure 5D in the Drop-seq paper by Macosko et al.
}
\examples{
vlnGrid(pbmc, head(rownames(pbmc@data), 15))
}
\author{
Selin Jessa
}
