% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whichCells}
\alias{whichCells}
\title{Retrieve cells in specified clusters}
\usage{
whichCells(seurat, clusters)
}
\arguments{
\item{seurat}{Seurat object}

\item{clusters}{Vector specifying clusters for which cells should be retrieved.
Elements should correspond to levels in \code{seurat@ident}.}
}
\value{
Character vector of cell names
}
\description{
Given a Seurat object and the names of specific clusters (corresponding
to identities in \code{seurat@ident}), return the names of cells
in those clusters.
}
\examples{
whichCells(pbmc, clusters = c(0, 1))
}
\author{
Selin Jessa
}
