% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addEmbedding}
\alias{addEmbedding}
\title{addEmbedding}
\usage{
addEmbedding(seurat, df, reduction = "tsne")
}
\arguments{
\item{seurat}{Seurat object, where dimensionality reduction has been applied,
i.e. (after applying Seurat::RunPCA() or Seurat::RunTSNE() to the object).}

\item{df}{Data frame with at least one column, "Cell", giving the cell ID}

\item{reduction}{String specifying the dimensionality reduction to use,
retrieves t-SNE by default. This should match the names of the elements of
the list seurat@dr, so it will typically be one of "pca" or "tsne".
Default: "tsne"}
}
\description{
Given a Seurat seurat and a data frame where the rows correspond to cells,
in the same order as in the Seurat seurat, add two columns giving coordinates
in a dimensionality reduced space.
}
\examples{
df <- data.frame(Cell = rownames(pbmc@meta.data),
                 Cluster = pbmc@meta.data$res.1)

addEmbedding(pbmc, df, reduction = "tsne")
}
\author{
Selin Jessa
}
