% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{clusterCenters}
\alias{clusterCenters}
\title{clusterCenters}
\usage{
clusterCenters(seurat)
}
\arguments{
\item{seurat}{Seurat object, where dimensionality reduction has been applied,
i.e. (after applying Seurat::RunTSNE() to the object).}
}
\value{
Data frame with three columns: Cluster, mean_tSNE_1, and mean_tSNE_2
}
\description{
Get centers of clusters given a Seurat object, to use for labelling
in tSNE space.
}
\examples{

clusterCenters(pbmc)
}
\author{
Selin Jessa
}
