% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytobox.R
\docType{package}
\name{cytobox}
\alias{cytobox}
\alias{cytobox-package}
\title{cytobox: A toolbox for analyzing single cell RNA-seq data}
\description{
Functions for processing, interrogating, and
visualizing single cell RNA-seq data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fungenomics/cytobox}
}

}
\author{
\strong{Maintainer}: Selin Jessa \email{selin.jessa@mail.mcgill.ca}

Authors:
\itemize{
  \item Alexis Blanchet-Cohen
  \item Yixing Hu
  \item Claudia L. Kleinman
}

Other contributors:
\itemize{
  \item Kleinman Lab, McGill University [contributor]
}

}
