% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{filterGenesSample}
\alias{filterGenesSample}
\title{filterGenesSample}
\usage{
filterGenesSample(genes, seurat, from_sp = "hg", to_sp = "mm")
}
\arguments{
\item{genes}{Character vector of genes to search for}

\item{seurat}{Seurat object}

\item{from_sp}{One of "hg" or "mm". Default: "hg"}

\item{to_sp}{One of "hg" or "mm". Default: "mm"}
}
\value{
Character vector of converted gene names which are detected in the sample
}
\description{
Given a set of genes, filter to the ones found in the sample.
This is useful because some Seurat functions (Seurat::VlnPlot, Seurat::FeaturePlot)
will quit on an error if a gene is not found (i.e. not in the rownames of seurat@data).
}
\examples{
filterGenesSample(c("IL32", "foo"), pbmc, "hg", "hg")
}
\author{
adapted from Claudia Kleinman
}
