% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_marker_selection.R
\name{getAllMarkers}
\alias{getAllMarkers}
\title{Title}
\usage{
getAllMarkers(df, labels, specify_clusters = NULL,
  output_graphs = FALSE, n_genes = 2, topn_markers = 10,
  graph_name = "Feature plot for cluster ", n_trees = NCOL(df) *
  n_genes, method = "RF", seurat)
}
\arguments{
\item{df}{A dataframe with the genes as columns and cells as rows}

\item{labels}{A dataframe that contains labels for the clustering solution. The dataframe should only have 1 column}

\item{specify_clusters}{If users only want to extract markers for a subset of clusters, enter the subset here in the form of a vector}

\item{output_graphs}{Indicates whether or not you wish to output the feature plots of the markers found to a pdf file}

\item{n_genes}{The number of genes used in each decision tree}

\item{topn_markers}{An integer indicating how many markers you wish to return for each cluster}

\item{graph_name}{The name of the feature plot if one wishes to output them to a pdf file}

\item{n_trees}{The number of trees one wishes to construct for each cluster}

\item{method}{the method used for extracting markers, currently only the randomForest(RF) method is avalible}
}
\value{
A dataframe that contains the list of markers and their associated attributes
}
\description{
Title
}
\examples{
data <- t(pbmc@scale.data)
labels <- data.frame(as.numeric(pbmc@meta.data$res.1))
marker_list<-getAllMarkers(data, labels = labels)

}
