% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{hg2mm}
\alias{hg2mm}
\title{hg2mm}
\usage{
hg2mm(...)
}
\arguments{
\item{...}{Human genes to convert, as individual strings or one character vector}
}
\value{
Character vector of converted names
}
\description{
Convert genes from human to mouse, wrapper around \code{\link{convertName}}
}
\examples{
hg2mm("TUBB", "NES")
hg2mm(c("TUBB", "NES"))
}
\author{
adapted from Claudia Kleinman
}
