% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelTSNEClusters.R
\name{labelTSNEClusters}
\alias{labelTSNEClusters}
\title{labelTSNEClusters}
\usage{
labelTSNEClusters(seurat, labelsFile)
}
\arguments{
\item{seurat}{Seurat object}

\item{labelsFile}{Path to labels file}
}
\value{
tSNE plot with labelled clusters
}
\description{
labelTSNEClusters
}
\author{
Alexis Blanchet-Cohen
}
