% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelTSNEClusters.R
\name{labelTSNEClusters}
\alias{labelTSNEClusters}
\title{labelTSNEClusters}
\usage{
labelTSNEClusters(seurat, labels = NULL, labelsOnPlot = TRUE,
  legend = TRUE)
}
\arguments{
\item{seurat}{Seurat object.}

\item{labels}{An data frame with 3 columns: "cluster", "label" and "colour". The "colour" column is optional.}

\item{labelsOnPlot}{Boolean to indicate whether labels should be displayed on tSNE plot.}

\item{legend}{Boolean to indicate whether legend should be plotted.}
}
\value{
A ggplot2 object. A tSNE plot with labelled clusters
}
\description{
labelTSNEClusters
}
\author{
Alexis Blanchet-Cohen
}
