% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelTSNEClusters.R
\name{labelTSNEClusters}
\alias{labelTSNEClusters}
\title{labelTSNEClusters}
\usage{
labelTSNEClusters(seurat, labels)
}
\arguments{
\item{seurat}{Seurat object.}

\item{labels}{A data frame with 3 columns: "cluster", "label" and "color". The "color" column is optional.}

\item{colors}{Cluster colors (Optional).}
}
\value{
A ggplot2 object. A tSNE plot with labelled clusters
}
\description{
labelTSNEClusters
}
\author{
Alexis Blanchet-Cohen
}
