% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{mm2hg}
\alias{mm2hg}
\title{mm2hg}
\usage{
mm2hg(x, ...)
}
\arguments{
\item{x}{Character vector or matrix with mouse gene symbols to convert.
If character, provide either individual strings or one character vector.
If a matrix, rownames should be gene symbols.}
}
\description{
Convert genes from mouse to human, wrapper around \code{\link{convertGenesBetweenSpeciesHomologs}}.
For vectors, the output is \strong{not order preserving}.
For matrices, order is preserved, but \strong{genes where the homolog is not
found are silently dropped}.
}
\examples{
mm2hg("Tubb5", "Nes")
mm2hg(c("Tubb5", "Nes"))

mat1 <- matrix(seq(1, 15), nrow = 3)
rownames(mat1) <- c("Gfap", "Fabp7", "Foo")
mm2hg(mat1)
}
\author{
Selin Jessa, adapted from Claudia Kleinman
}
