% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{mm2hg}
\alias{mm2hg}
\title{mm2hg}
\usage{
mm2hg(...)
}
\arguments{
\item{...}{Mouse genes to convert, as individual strings or one character vector}
}
\value{
Character vector of converted names
}
\description{
Wrapper around \code{\link{convertName}}
}
\examples{
mm2hg("Tubb5", "Nes")
}
