% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plotting.R
\name{theme_min}
\alias{theme_min}
\title{Apply a clean theme to a ggplot2 object}
\usage{
theme_min(base_size = 11, base_family = "")
}
\description{
Apply a clean theme to a ggplot2 object
}
\references{
https://github.com/sjessa/ggmin
}
\author{
Selin Jessa
}
