% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tsne}
\alias{tsne}
\title{tsne}
\usage{
tsne(seurat, colour_by = NULL, colours = NULL, label = TRUE,
  point_size = 0.6, alpha = 0.8, legend = ifelse((is.null(colour_by)) &&
  (label), FALSE, TRUE), label_repel = TRUE, label_size = 4, title = NULL)
}
\arguments{
\item{seurat}{Seurat object, where Seurat::RunTSNE() has been applied}

\item{colour_by}{(Optional) String, specifying the column in \code{seurat@meta.data}
by which to colour cells. Default: NULL, colour cells by cluster (in \code{seurat@ident}).}

\item{colours}{(Optional) Named character vector of colours for points. Names should
correspond to cluster names (e.g. \code{levels(seurat@ident)}), or to categorical
values in the column of \code{seurat@meta.data} specified in \code{colour_by}. Default:
use default ggplot2 colours.}

\item{label}{Logical, whether to plot cluster labels. Default: TRUE}

\item{point_size}{Numeric, size of points in scatter plot. Default: 0.6}

\item{alpha}{Numeric, fixed alpha value for points: Default: 0.8}

\item{legend}{Logical, whether to plot legend. Default: FALSE if \code{colour_by}
is NULL and \code{label} is TRUE, true otherwise.}

\item{label_repel}{Logical, if \code{label} is TRUE, whether to plot cluster
labels repelled from the center, on a slightly transparent white background and
with an arrow pointing to the cluster center. If FALSE, simply plot the
cluster label at the cluster center. Default: TRUE.}

\item{label_size}{Numeric, controls the size of text labels. Default: 4.}

\item{title}{(Optional) String specifying title.}
}
\value{
A ggplot2 object
}
\description{
Plot a coloured, labelled tSNE plot for a datasets, akin to Seurat::TSNEPlot()
}
\examples{
tsne(pbmc)
}
\author{
Selin Jessa
}
