% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computed_plots.R
\name{tsneByMeanMarkerExpression}
\alias{tsneByMeanMarkerExpression}
\title{Colour cells in t-SNE space by gene expression}
\usage{
tsneByMeanMarkerExpression(seurat, genes, reduction = "tsne",
  label = TRUE, palette = "redgrey", title = NULL, alpha = 0.6,
  label_repel = TRUE, label_size = 4, legend = TRUE,
  hide_ticks = FALSE, limits = NULL, label_short = FALSE)
}
\arguments{
\item{seurat}{Seurat object, where dimensionality reduction has been applied,
i.e. (after applying Seurat::RunPCA() or Seurat::RunTSNE() to the object)}

\item{genes}{String or character vector specifying gene(s) to use}

\item{reduction}{String specifying the dimensionality reduction to use,
retrieves t-SNE by default. This should match the names of the elements of
the list seurat@dr, so it will typically be one of "pca" or "tsne".
Default: "tsne"}

\item{label}{Logical, whether to label clusters on the plot. Default: TRUE.}

\item{palette}{String or character vector. If a string,
one of "viridis", "blues", or "redgrey", specifying which gradient
palette to use. Otherwise, a character vector of colours (from low to high)
to interpolate to create the scael. Default: redgrey.}

\item{title}{(Optional) String specifying the plot title}

\item{alpha}{Numeric, fixed alpha for points. Default: 0.6}

\item{label_repel}{Logical, if \code{label} is TRUE, whether to plot cluster
labels repelled from the center, on a slightly transparent white background and
with an arrow pointing to the cluster center. If FALSE, simply plot the
cluster label at the cluster center. Default: TRUE.}

\item{label_size}{Numeric, controls the size of text labels. Default: 4.}

\item{legend}{Logical, whether or not to plot legend. Default: TRUE}

\item{hide_ticks}{Logical, whether to hide axis ticks. Default: FALSE}

\item{limits}{(Optional) A numeric vector of length two providing the limits to
use for the colour scale (documentation
from \code{\link[ggplot2]{continous_scale}}. Default: 0 and max of the data.}

\item{label_short}{(Optional/Experimental!!) Logical, if TRUE, assumes clusters
(at seurat@ident) consist of a prefix and a suffix separated by a non-alpha
numeric character (\code{"[^[:alnum:]]+"}), and tries to separate these names
and only plot the prefix, for shorter labels and a cleaner plot. Default: FALSE.}
}
\value{
A ggplot object
}
\description{
Plot a low-dimensional embedding of the cells,
coloured by expression of a gene, or mean expression of a group of marker
genes. Defaults to t-SNE space, but see the \code{reduction} argument for
how to plot in PCA space instead.
}
\examples{
tsneByMeanMarkerExpression(pbmc, "IL32")
tsneByMeanMarkerExpression(pbmc, c("IL32", "CD2"), reduction = "pca")
tsneByMeanMarkerExpression(pbmc, c("IL32", "CD2"), reduction = "pca")
}
\author{
Selin Jessa
}
