% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tsneByMeanMarkerExpression}
\alias{tsneByMeanMarkerExpression}
\title{tsneByMeanMarkerExpression}
\usage{
tsneByMeanMarkerExpression(seurat, genes, reduction = "tsne", title = NULL,
  legend = TRUE, hide_ticks = FALSE)
}
\arguments{
\item{seurat}{Seurat object, where dimensionality reduction has been applied,
i.e. (after applying Seurat::RunPCA() or Seurat::RunTSNE() to the object)}

\item{genes}{String or character vector specifying gene(s) to use}

\item{reduction}{String specifying the dimensionality reduction to use,
retrieves t-SNE by default. This should match the names of the elements of
the list seurat@dr, so it will typically be one of "pca" or "tsne".
Default: "tsne"}

\item{title}{(Optional) String specifying the plot title}

\item{legend}{Logical, whether or not to plot legend. Default: TRUE}

\item{hide_ticks}{Logical, whether to hide axis ticks. Default: FALSE}
}
\value{
A ggplot object
}
\description{
Plot a low-dimensional embedding of the cells,
coloured by expression of a gene, or mean expression of a group of marker
genes. Defaults to t-SNE space, but see the \code{reduction} argument for
how to plot in PCA space instead.
}
\author{
Selin Jessa
}
