% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tsneByPercentileMarkerExpression}
\alias{tsneByPercentileMarkerExpression}
\alias{dashboard}
\title{tsneByPercentileMarkerExpression}
\usage{
tsneByPercentileMarkerExpression(seurat, genes, label = TRUE,
  reduction = "tsne", title = NULL, palette = "viridis", extra = FALSE,
  verbose = FALSE)
}
\arguments{
\item{seurat}{Seurat object, where dimensionality reduction has been applied,
i.e. (after applying Seurat::RunPCA() or Seurat::RunTSNE() to the object)}

\item{genes}{String or character vector specifying gene(s) to use}

\item{label}{Logical, whether to label clusters on the plot. Default: TRUE.}

\item{reduction}{String specifying the dimensionality reduction to use,
retrieves t-SNE by default. This should match the names of the elements of
the list seurat@dr, so it will typically be one of "pca" or "tsne".
Default: "tsne"}

\item{title}{(Optional) string used as title for the plot.}

\item{palette}{String, one of "viridis" or "blues", specifying which gradient
palette to use. Default: viridis.}

\item{extra}{Logical, plot a detailed three-panel plot, where the first
is a proportional bar plot of cells in each cluster in each percentile
group, the second is a ridge plot showing density in each cluster of the mean
expression of the markers (coloured by median percentile group within the
cluster), and the third is the labelled tSNE plot coloured
by percentiles. Requires \code{label = TRUE}. Default: FALSE.
Default: FALSE.}

\item{verbose}{Logical, whether to print status updates. Default: FALSE.}
}
\value{
A ggplot object
}
\description{
Plot a low-dimensional embedding of the cells,
coloured by expression percentile of a gene, of the total expression of a
group of marker genes.
}
\examples{
tsneByPercentileMarkerExpression(pbmc, "IL32")
tsneByPercentileMarkerExpression(pbmc, c("IL32", "CD2"), reduction = "pca")
dashboard(pbmc, "IL32", "Test dashboard")
}
\author{
Selin Jessa
}
