% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tsneByPercentileMarkerExpression}
\alias{tsneByPercentileMarkerExpression}
\title{tsneByPercentileMarkerExpression}
\usage{
tsneByPercentileMarkerExpression(seurat, markers, label = TRUE,
  reduction = "tsne", palette = "viridis", extra = FALSE)
}
\arguments{
\item{seurat}{Seurat object, where dimensionality reduction has been applied,
i.e. (after applying Seurat::RunPCA() or Seurat::RunTSNE() to the object)}

\item{markers}{String or character vector specifying gene(s) to use}

\item{label}{Logical, whether to label clusters on the plot. Default: TRUE.}

\item{reduction}{String specifying the dimensionality reduction to use,
retrieves t-SNE by default. This should match the names of the elements of
the list seurat@dr, so it will typically be one of "pca" or "tsne".
Default: "tsne"}

\item{palette}{String, one of "viridis" or "blues", specifying which gradient
palette to use. Default: viridis.}
}
\value{
A ggplot object
}
\description{
Plot a low-dimensional embedding of the cells,
coloured by expression percentile of a gene, of the total expression of a
group of marker genes.
}
\examples{
tsneByPercentileMarkerExpression(pbmc, "IL32")
tsneByPercentileMarkerExpression(pbmc, c("IL32", "CD2"), reduction = "pca")
}
\author{
Selin Jessa
}
