% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vln}
\alias{vln}
\title{vln}
\usage{
vln(seurat, genes, facet_by = "cluster")
}
\arguments{
\item{seurat}{Seurat object}

\item{genes}{Genes to plot violins for}

\item{facet_by}{String, one of "gene" or "cluster". Default: "cluster",
genes will be on the x axis, and there will be one plot per cluster.
If "gene", clusters will be on the x axis, and there will be one plot per
gene (akin to Seurat::VlnPlot)}
}
\value{
A ggplot2 object
}
\description{
Similar to Seurat::VlnPlot() except it silently skips genes that are not found in the
data, and has an option to group plots by cluster instead of gene.
}
\examples{
vln(pbmc, c("IL32", "MS4A1"))
vln(pbmc, c("IL32", "MS4A1"), facet_by = "gene")
}
\author{
Selin Jessa
}
